+++ backend/chem/urls.py
@@ -1,8 +1,8 @@
-from rest_framework.routers import DefaultRouter
+from config.routers import NumericPkRouter
 
 from .views import ChemDrawingViewSet
 
-router = DefaultRouter()
+router = NumericPkRouter()
 router.register('chem-drawings', ChemDrawingViewSet, basename='chem-drawing')
 
 urlpatterns = router.urls